requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
supabase==2.0.0
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
pathlib==1.0.1
